version 0.2

+++ src/centre/centre.controller.ts
@@ -1,8 +1,10 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Request } from '@nestjs/common';
 import { CentreService } from './centre.service';
 import { CreateCentreDto } from './dto/create-centre.dto';
 import { UpdateCentreDto } from './dto/update-centre.dto';
 import { UserGuard } from 'src/users/Guards/user.guard';
+import { User } from 'src/users/Decorators/User';
+
 
 @Controller('centre')
 @UseGuards(UserGuard)
@@ -17,6 +19,7 @@ export class CentreController {
   @Get()
   findAll() {
     return this.centreService.findAll();
+    
   }
 
   @Get(':id')
@@ -25,13 +28,15 @@ export class CentreController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateCentreDto: UpdateCentreDto) {
-    return this.centreService.update(+id, updateCentreDto);
+  update(@Param('id') id: string, @Body() updateCentreDto: UpdateCentreDto,@User() user: any) {
+    if (user.role==='admin')
+      return this.centreService.update(+id, updateCentreDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.centreService.remove(+id);
+  remove(@Param('id') id: string,@User() user: any) {
+    if (user.role==='admin')
+      return this.centreService.remove(+id);
   }
   @Get('centres/:villeId')
   findByRegion(@Param('villeId') villeId: number) {
